refactored code + added 1el oxidation algorithm (NOT FINAL)
