+++ setup.py
@@ -3,7 +3,7 @@ import setuptools
 
 setuptools.setup(
     name="GES-comp-echem",
-    version="0.1.23a",
+    version="0.1.24a",
     description="",
     long_description="",
     packages=["compechem"],
